requirements.txt
Django==2.0
gunicorn==19.6.0
django-admin-bootstrapped==2.5.7
django-background-tasks==1.2.0
django-colorfield==0.1.15
django-compat==1.0.15
django-cors-headers==3.2.1
django-crispy-forms==1.8.1
django-datetimepicker==3.14
django-grappelli==2.13.1
django-import-export==1.2.0
django-jet==1.0.8
whitenoise==3.2
django-markdown==0.8.4
django-qr-code==1.0.0
django-smart-selects==1.5.4
django-smart-selects-711==1.0.0
django-templated-mail==1.1.1
djangorestframework==3.11.0
djangorestframework-jwt==1.11.0
djangorestframework-simplejwt==4.4.0